requirements.txt
opencv-python ~= 4.8.1.78
Pillow ~= 10.1.0
numpy ~= 1.26.2
matplotlib ~= 3.8.2
